Update index.html
In a rush but here's an update. MeiliSearch is now included. There is some untested code for grabbing MeiliSearch keys and automatically importing them into index.html for you. This may or may not work. The .env file is finished however it is not automatically adjusting the container names in docker-compose.yml just yet. Wait 24+- hours or so for a completed update.