Update app.py
Changed the page numbers as my public version was getting too long (61 pages). There is a configurable option found in the .env called "pagenumdisplaylimit" and the default is set to 15. From the main page you will see pages 1-15 and "Next" for forwarding a single page, "Forwards ➡️" which will forward the number assiged by "pagenumdisplaylimit", and "Last" to skip to the last page.
Likewise at page 15, you will see pages 8-22 and see a "First", "Backwards ⬅️", and "Previous" to do the same thing but in the opposite direction.